// pages/interface.tsx
export interface Topic {
  id: number;
  name: string;
}

export interface Author {
  id: number;
  full_name: string;
  email: string;
  avatar_url?: string;
}

export interface User {
  id: number;
  full_name: string;
  email: string;
}

export interface Comment {
  id: number;
  content: string;
  user: User;
  created_at: string;
}

export interface Post {
  id: number;
  title: string;
  slug: string;
  content: string;
  short_description?: string;
  thumbnail?: string;
  published_at: string;
  topics?: Topic[];
  author?: Author;
  postComments?: Comment[];
}

export interface ApiResponse<T> {
  data: T;
  message?: string;
}

export interface PostsResponse {
  posts: Post[];
  total: number;
}